refactor: rename utils file and rename imports
