+++ src/utils/addVariablesValues.ts
